Working on adjusting carousel code and being able to reuse any array
